directory structure of project added

+++ src/main/java/com/springboot/stackoverflow/entity/Comment.java
@@ -13,7 +13,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
     @Column(name = "email")
     private String email;
 
@@ -76,11 +76,11 @@ public class Comment {
         this.answer = answer;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
